+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 20.11.1
-  * Lessened push notification security restrictions 
+  * Lessened push notification security restrictions
+  * Added a default way to acquire app start timestamp for APM
+  * Added a way to override the current app start timestamp for APM
+  * Added manual trigger for when app has finished loading for APM
 
 ## 20.11.0
   * !! Consent change !! To record orientation you now need to give "user" consent